react project/backend/requirements.txt
numpy 
pandas 
librosa 
wave
pyrebase4